Docs/conf.py: Explicitly define external URL schemes
By default, any URL with a scheme of "http", "https", "ftp", or "mailto"
is treated as an external link. Since the "ircs" scheme is not included,
the IRC link in community/forums.md does not get resolved as an external
link, and instead tries to link to a header in the docs themselves. Fix
this by explicitly defining which schemes should resolve to external
links using the myst_url_schemes configuration option [1], which is now
set to the default schemes along with "ircs".

This fixes the "cross-reference target not found" warning for
'ircs://irc.libera.chat/#coreboot'

[1] https://myst-parser.readthedocs.io/en/latest/syntax/cross-referencing.html#customising-external-url-resolution

Change-Id: I9e1c76b2bacbacaa06340f940c76b50de38e43e8
Signed-off-by: Nicholas Chin <nic.c3.14@gmail.com>
Reviewed-on: https://review.coreboot.org/c/coreboot/+/84069
Reviewed-by: Nico Huber <nico.h@gmx.de>
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>