mb/intel/coffeelake_rvp/cml_u: Correct number of jacks in hda_verb.c
This was found due to the `_Static_assert()` from CB:84360 failing.

As I do not own this board, I cannot test whether is change is
"functionally correct". However, I believe it is more likely that the
original authors forgot to update the verb table size, rather than them
adding additional verb data which was not meant to be used.

TEST=`_Static_assert()` mentioned above does not fail anymore.

Change-Id: I8df44e056bc841bfb344749ba214e6fb71a1955b
Signed-off-by: Nicholas Sudsgaard <devel+coreboot@nsudsgaard.com>
Reviewed-on: https://review.coreboot.org/c/coreboot/+/84487
Reviewed-by: Felix Singer <service+coreboot-gerrit@felixsinger.de>
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>