soc/intel/cmn/blk: cse_enable_ptt: Wait up to 5 s for FW Init Complete
FW Init Complete is a prerequisite for sending the FW FEATURE SHIPMENT
TIME STATE OVERRIDE message. Unfortunately, on some platforms such as
Lenovo ThinkCentre M700 Tiny, it takes too long for the flag to be set,
so enabling PTT fails.

Wait up to 5 seconds for the FW Init Complete to be set instead of
failing immediately.

On M700 Tiny with debug level set to ERROR, we have to wait nearly 2
seconds:

    [EMERG]  HECI: CSE took 1900 ms to complete FW init

Because FW Init Complete is not required for getting the current feature
enablement state, only for setting, move the FW Init Complete check to
after we've determined if we actually need to change the state. This
avoids needlessly increasing boot time.

Reference: Intel ME 11.x BIOS Specification, #549522, section 6.3.15

Change-Id: Ib6de170f3f998273bec437848faa49652f013f45
Signed-off-by: Michał Kopeć <michal.kopec@3mdeb.com>
Reviewed-on: https://review.coreboot.org/c/coreboot/+/84862
Reviewed-by: Martin L Roth <gaumless@gmail.com>
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>